### 2023-08-08 #128 @AlexanderJuestel
GemGIS - Spatial Data Processing for Geomodeling
Submitting Author: Name (@AlexanderJuestel)
Package Name: GemGIS
One-Line Description of Package: Spatial Data Processing for Geomodeling
Repository Link:  https://github.com/cgre-aachen/gemgis
Version submitted:  1.0.11 (soon 1.1 with bug fixes for some methods and the API Reference once it is working, no major functionality changes)
Documentation: https://gemgis.readthedocs.io/en/latest/
JOSS Publication: https://joss.theoj.org/papers/10.21105/joss.03709
JOSE Publication (awaiting publication): https://github.com/openjournals/jose-reviews/issues/185
Version accepted: TBD

---

GemGIS is a Python-based, open-source geographic information processing library. It is capable of preprocessing spatial data such as vector data (shape files, geojson files, geopackages,…), raster data (tif, png,…), data obtained from online services (WCS, WMS, WFS) or XML/KML files (soon). Preprocessed data can be stored in a dedicated Data Class to be passed to the geomodeling package [GemPy](https://github.com/cgre-aachen/gempy) in order to accelerate the model building process. Postprocessing of model results will allow export from GemPy to geoinformation systems such as QGIS and ArcGIS or to Google Earth for further use.

GemGIS uses and combines the full functionality of [GeoPandas](https://geopandas.org/), [rasterio](https://rasterio.readthedocs.io/en/latest/), [OWSLib](https://geopython.github.io/OWSLib/), [Pandas](https://pandas.pydata.org/docs/), [Shapely](https://shapely.readthedocs.io/en/latest/manual.html), [PyVista](https://docs.pyvista.org/) and [NumPy](https://numpy.org/) to simplify, accelerate and automate the workflows used to preprocess spatial data for geomodeling.

In addition, [almost 70 tutorials](https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html) illustrate the different functionalities of GemGIS.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

**Data extraction/processing:** GemGIS uses the functionality of packages like GeoPandas, Shapely, or Rasterio to extract information from vector and raster data and to put it in a form that it can be used by the [GemPy](https://github.com/cgre-aachen/gempy).

**Data visualization:** GemGIS uses the functionality of packages like matplotlib or PyVista to create static and dynamic plots of data and meshes. This includes digital elevation models or meshes of subsurfaces layers, boreholes geological cross sections or even seismic data.

**Workflow automation:** The entire purpose of GemGIS is to provide methods to accelerate the preparation of input data for GemPy. Over time, the package has also gained additional functionality to work with a variety of datasets utilized for subsurface applications, see [Tutorials](https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html).

The target audience is the open-source Geosciences community, researchers, students but also industry. GemGIS provides functionality to accelerate the preparation of input data for the structural geological modeling package GemPy which has been used in numerous publications. For applications at universities, we are in the final stages of getting a [JOSE publication](https://github.com/openjournals/jose-reviews/issues/185#issuecomment-1549230502) approved with more than 20 structural geological models that are used at RWTH Aachen University for teaching purposes and where GemGIS and GemPy will be included in future courses.

GemGIS does not reinvent the wheel but rather combines the functionality of already existing packages mentioned in the description above. The packages utilized the most in GemGIS are the well-known packages like GeoPandas, Shapely, Rasterio, Pandas, NumPy, PyVista, matplotlib, etc. We also decided against i.e. wrapping GeoPandas GeoDataFrames in our own class or creating many new classes so that users can still use the full functionality of the underlying packages. This is one big advantage in comparison to GemPy where i.e. the meshes of the resulting structural geological models cannot be extracted (GemGIS is capable of extracting them though). Another example is that raster data opened with GemGIS will be stored as PyVista PolyData datasets or as grids so that users can harvest the functionality of this amazing package.

  - If you made a pre-submission inquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted:

Pre-submission inquiry: #126, @NickleDave

- [ ] contains a README with instructions for installing the development version. **--> will be checked once added to the Readme**
- [ ] includes documentation with examples for all functions. **--> will be checked once it is working again**

- [ ] Do you wish to automatically submit to the [Journal of Open Source Software][JournalOfOpenSourceSoftware]? No, JOSS publication already existing: https://joss.theoj.org/papers/10.21105/joss.03709
