### 2023-08-08 #128 @AlexanderJuestel
Hello,
There are two bullets that I still need to check, I am further open to make updates to the package to meet minimal criteria before review (Section 5 of the Maintainer Guide) if there is still anything missing.
