+++ README.md
@@ -1,6 +1,6 @@
 <p align="center"><img src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/docs/getting_started/images/Modern1.png" width="600">
 
-> Spatial data and information processing for geomodeling
+> Spatial data and information processing for geomodeling and subsurface data
 
 
 [![PyPI](https://img.shields.io/badge/python-3-blue.svg)](https://www.python.org/downloads/)
@@ -17,11 +17,11 @@
 
 ## Overview 
 
-We attempt to simplify the access to open-source spatial data processing for geological modeling with the development of **GemGIS, a Python-based open-source library**. 
+We attempt to simplify the access to open-source spatial data processing for geological modeling and for subsurface data (wells, seismic, etc.) with the development of **GemGIS, a Python-based open-source library**. 
 
 GemGIS wraps and extends the functionality of packages known to the geo-community such as [GeoPandas](https://geopandas.org/), [rasterio](https://rasterio.readthedocs.io/en/latest/#), [OWSLib](https://geopython.github.io/OWSLib/), [Shapely](https://shapely.readthedocs.io/en/latest/manual.html), [PyVista](https://docs.pyvista.org/), [Pandas](https://pandas.pydata.org/), [NumPy](https://numpy.org/) and the geomodeling package [GemPy](https://docs.gempy.org/). 
 
-The aim of GemGIS, as indicated by the name, is to become a bridge between conventional geoinformation systems (GIS) such as ArcGIS and QGIS, and geomodeling tools such as GemPy, allowing simpler and more automated workflows from one environment to the other. This also includes making it simpler to visualize the results obtained from GemGIS and GemPy with PyVista or Blender. 
+The aim of GemGIS, as indicated by the name, is to become a bridge between conventional geoinformation systems (GIS) such as ArcGIS and QGIS, and geomodeling tools such as GemPy, allowing simpler and more automated workflows from one environment to the other. This also includes making it simpler to visualize the results obtained from GemGIS and GemPy with PyVista or Blender. Further, subsurface data processing workflows are implemented for integrated data analyses.
 
 <p align="center"><img src="https://raw.githubusercontent.com/cgre-aachen/gemgis/main/joss/images/fig1.png" width="800">
 
@@ -31,7 +31,7 @@ The aim of GemGIS, as indicated by the name, is to become a bridge between conve
 [Find the documentation of GemGIS here](https://gemgis.readthedocs.io/en/latest/index.html). It includes introductions to the main libraries used and to introductory topics like "What is vector data?" or "What is raster data?". 
 
 In addition, [tutorial notebooks](https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html) provide an overview of the different features of GemGIS. The notebooks can also be downloaded directly from [here](https://rwth-aachen.sciebo.de/s/AfXRsZywYDbUF34/download?path=%2F&files=tutorials01_53.zip).
-Furthermore, many [example models](https://gemgis.readthedocs.io/en/latest/getting_started/example/index.html) showcase a variety of geological structures that can be modeled with GemGIS and GemPy.
+Furthermore, many [example models](https://gemgis.readthedocs.io/en/latest/getting_started/example/index.html) showcase a variety of geological structures that can be modeled with GemGIS and GemPy. The [API Reference](https://gemgis.readthedocs.io/en/latest/getting_started/api.html#) provides an overview of the implemented functions in GemGIS and how to use them. 
 
 
 
@@ -44,6 +44,7 @@ It is recommended to use GemGIS with **python">=3.10"** in a separated environme
 
[…]
