+++ environment_dev.yml
@@ -8,9 +8,10 @@ dependencies:
   #- pyvista=0.39.1
   - pip
   - pip:
+    - gempy==2.3
     - rasterio==1.3.8
     - geopandas==0.13.2
-    - pyvista== 0.41.1
+    - pyvista
     - shapely
     - fiona
     - pyproj
@@ -20,7 +21,6 @@ dependencies:
     - scipy
     - pooch<1.5.0
     - gemgis==1.0.11
-    - gempy==2.3
     - pyqt5
     - nbsphinx
     - sphinx_markdown_tables
