+++ docs/conf.py
@@ -25,6 +25,7 @@ import os
 import sys
 sys.path.insert(0, os.path.abspath(".."))
 sys.path.insert(0, os.path.abspath('../..'))
+sys.path.insert(0, os.path.abspath("gemgis/"))
 sys.path.insert(0, os.path.abspath("../gemgis/"))
 sys.path.insert(0, os.path.abspath("../../gemgis/"))
 
