### 2023-08-15 ac0f4d8
Fixed issue found in https://github.com/cgre-aachen/gemgis/issues/294
