### 2023-08-15 #128 @NickleDave
Not sure if I spotted the issue but I replied there: https://stackoverflow.com/a/76908429/4906855

You can also feel free to ask for help in our Discourse in the future (https://pyopensci.discourse.group/) (or now, if I haven't helped :sweat_smile:)
