+++ docs/getting_started/api.rst
@@ -6,424 +6,6 @@ GemGIS API Reference
 The API reference provides an overview of all functions and methods implemented in GemGIS.
 
 
-Vector
-______
-
-The following sections provide an overview of the methods implemented in the GemGIS Vector module.
-
-Extracting Coordinates from GeoDataFrames
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The following methods are used to extract X and Y coordinates or X, Y, and Z coordinates from GeoDataFrames or Shapely Base Geometries for further usage.
-
-.. autosummary::
-   :toctree: reference/vector_api
-
-   gemgis.vector.extract_xy
-
-   gemgis.vector.extract_xy_linestring
-   gemgis.vector.extract_xy_linestrings
-   gemgis.vector.extract_xy_points
-   gemgis.vector.extract_xyz
-   gemgis.vector.extract_xyz_array
-   gemgis.vector.extract_xyz_rasterio
-   gemgis.vector.extract_xyz_points
-   gemgis.vector.extract_xyz_linestrings
-   gemgis.vector.extract_xyz_polygons
-
-
-Extracting Coordinates and Intersections from GeoDataFrames/Shapely Polygons
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The following methods are used to extract coordinates and intersections from GeoDataFrames containing Shapely Polygons
-or from single Shapely Polygons or multiple Shapely Polygons.
-
-
-.. autosummary::
-   :toctree: reference/vector_api
-
-   gemgis.vector.extract_xy_from_polygon_intersections
-   gemgis.vector.intersection_polygon_polygon
-   gemgis.vector.intersections_polygon_polygons
-   gemgis.vector.intersections_polygons_polygons
-
-
-Calculating and extracting coordinates from cross sections
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The following methods are used to calculate and extract coordinates from cross sections where data was stored as Shape
-Files and is now loaded as GeoDataFrames.
-
-.. autosummary::
-   :toctree: reference/vector_api
-
-   gemgis.vector.extract_interfaces_coordinates_from_cross_section
-   gemgis.vector.extract_xyz_from_cross_sections
-   gemgis.vector.calculate_coordinates_for_point_on_cross_section
-   gemgis.vector.calculate_coordinates_for_linestring_on_cross_sections
-   gemgis.vector.calculate_coordinates_for_linestrings_on_cross_sections
-   gemgis.vector.extract_interfaces_coordinates_from_cross_section
-
-Calculating and extracting angles from cross sections and maps
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-The following methods are used to calculate and extract angles from cross sections.
-
-.. autosummary::
-   :toctree: reference/vector_api
-
-   gemgis.vector.calculate_angle
-   gemgis.vector.calculate_azimuth
-   gemgis.vector.calculate_strike_direction_straight_linestring
-   gemgis.vector.calculate_strike_direction_bent_linestring
-   gemgis.vector.calculate_dipping_angle_linestring
-   gemgis.vector.calculate_dipping_angles_linestrings
-   gemgis.vector.calculate_orientation_from_bent_cross_section
-   gemgis.vector.calculate_orientation_from_cross_section
-   gemgis.vector.calculate_orientations_from_cross_section
[…]
