+++ docs/getting_started/reference/visualization_api/gemgis.visualization.create_lines_3d_polydata.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.create\_lines\_3d\_polydata
-================================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: create_lines_3d_polydata
