+++ docs/getting_started/reference/visualization_api/gemgis.visualization.create_polydata_from_dxf.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.create\_polydata\_from\_dxf
-================================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: create_polydata_from_dxf
