+++ docs/getting_started/reference/web_api/gemgis.web.load_as_file.rst
@@ -1,6 +1,0 @@
-﻿gemgis.web.load\_as\_file
-=========================
-
-.. currentmodule:: gemgis.web
-
-.. autofunction:: load_as_file
