+++ docs/getting_started/reference/web_api/gemgis.web.load_wms.rst
@@ -1,6 +1,0 @@
-﻿gemgis.web.load\_wms
-====================
-
-.. currentmodule:: gemgis.web
-
-.. autofunction:: load_wms
