+++ docs/getting_started/reference/vector_api/gemgis.vector.explode_multilinestrings.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.explode\_multilinestrings
-=======================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: explode_multilinestrings
