+++ docs/getting_started/reference/utils_api/gemgis.utils.set_resolution.rst
@@ -1,6 +1,0 @@
-﻿gemgis.utils.set\_resolution
-============================
-
-.. currentmodule:: gemgis.utils
-
-.. autofunction:: set_resolution
