+++ docs/getting_started/reference/postprocessing_api/gemgis.postprocessing.calculate_dip_and_azimuth_from_mesh.rst
@@ -1,6 +1,0 @@
-﻿gemgis.postprocessing.calculate\_dip\_and\_azimuth\_from\_mesh
-==============================================================
-
-.. currentmodule:: gemgis.postprocessing
-
-.. autofunction:: calculate_dip_and_azimuth_from_mesh
