+++ docs/getting_started/reference/vector_api/gemgis.vector.create_voronoi_polygons.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.create\_voronoi\_polygons
-=======================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: create_voronoi_polygons
