+++ docs/getting_started/reference/raster_api/gemgis.raster.save_as_tiff.rst
@@ -1,6 +1,0 @@
-﻿gemgis.raster.save\_as\_tiff
-============================
-
-.. currentmodule:: gemgis.raster
-
-.. autofunction:: save_as_tiff
