+++ docs/getting_started/reference/utils_api/gemgis.utils.get_locations.rst
@@ -1,6 +1,0 @@
-﻿gemgis.utils.get\_locations
-===========================
-
-.. currentmodule:: gemgis.utils
-
-.. autofunction:: get_locations
