+++ docs/getting_started/reference/vector_api/gemgis.vector.explode_multilinestring.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.explode\_multilinestring
-======================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: explode_multilinestring
