+++ docs/getting_started/reference/vector_api/gemgis.vector.create_linestring_from_points.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.create\_linestring\_from\_points
-==============================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: create_linestring_from_points
