+++ docs/getting_started/reference/raster_api/gemgis.raster.read_ts.rst
@@ -1,6 +1,0 @@
-﻿gemgis.raster.read\_ts
-======================
-
-.. currentmodule:: gemgis.raster
-
-.. autofunction:: read_ts
