+++ docs/getting_started/reference/utils_api/gemgis.utils.convert_to_gempy_df.rst
@@ -1,6 +1,0 @@
-﻿gemgis.utils.convert\_to\_gempy\_df
-===================================
-
-.. currentmodule:: gemgis.utils
-
-.. autofunction:: convert_to_gempy_df
