+++ docs/getting_started/reference/vector_api/gemgis.vector.extract_xy_linestring.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.extract\_xy\_linestring
-=====================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: extract_xy_linestring
