+++ docs/getting_started/reference/visualization_api/gemgis.visualization.get_petrel_cmap.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.get\_petrel\_cmap
-======================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: get_petrel_cmap
