+++ docs/getting_started/reference/vector_api/gemgis.vector.sort_by_stratigraphy.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.sort\_by\_stratigraphy
-====================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: sort_by_stratigraphy
