+++ docs/getting_started/reference/visualization_api/gemgis.visualization.create_borehole_tube.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.create\_borehole\_tube
-===========================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: create_borehole_tube
