+++ docs/getting_started/reference/visualization_api/gemgis.visualization.resample_between_well_deviation_points.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.resample\_between\_well\_deviation\_points
-===============================================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: resample_between_well_deviation_points
