+++ docs/getting_started/reference/vector_api/gemgis.vector.calculate_orientation_for_three_point_problem.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.calculate\_orientation\_for\_three\_point\_problem
-================================================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: calculate_orientation_for_three_point_problem
