+++ docs/getting_started/reference/vector_api/gemgis.vector.calculate_orientation_from_cross_section.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.calculate\_orientation\_from\_cross\_section
-==========================================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: calculate_orientation_from_cross_section
