+++ docs/getting_started/reference/vector_api/gemgis.vector.intersections_polygons_polygons.rst
@@ -1,6 +1,0 @@
-﻿gemgis.vector.intersections\_polygons\_polygons
-===============================================
-
-.. currentmodule:: gemgis.vector
-
-.. autofunction:: intersections_polygons_polygons
