+++ docs/getting_started/reference/visualization_api/gemgis.visualization.create_thickness_maps.rst
@@ -1,6 +1,0 @@
-﻿gemgis.visualization.create\_thickness\_maps
-============================================
-
-.. currentmodule:: gemgis.visualization
-
-.. autofunction:: create_thickness_maps
