+++ docs/getting_started/reference/raster_api/gemgis.raster.extract_contour_lines_from_raster.rst
@@ -1,6 +1,0 @@
-﻿gemgis.raster.extract\_contour\_lines\_from\_raster
-===================================================
-
-.. currentmodule:: gemgis.raster
-
-.. autofunction:: extract_contour_lines_from_raster
