+++ gemgis/postprocessing.py
@@ -21,9 +21,10 @@ GNU General Public License (LICENSE) for more details.
 
 from shapely import geometry
 import geopandas as gpd
+import pandas as pd
 import numpy as np
 from typing import List, Union
-from gemgis import gemgis
+from gemgis import gemgis, visualization
 import pyvista as pv
 import pyproj
 import xml
@@ -998,3 +999,331 @@ def save_qgis_qml_file(gdf: gpd.geodataframe.GeoDataFrame,
     tree.write(path, encoding="utf-8", xml_declaration=False)
 
     print('QML file successfully saved as %s' % path)
+
+
+def clip_fault_of_gempy_model(geo_model,
+                              fault: str,
+                              which: str = 'first',
+                              buffer_first: Union[int, float] = None,
+                              buffer_last: Union[int, float] = None,
+                              i_size: Union[int, float] = 1000,
+                              j_size: Union[int, float] = 1000,
+                              invert_first: bool = True,
+                              invert_last: bool = False) -> Union[
+    pv.core.pointset.PolyData, List[pv.core.pointset.PolyData]]:
+    """
+    Clip fault of a GemPy model.
+
+    Parameters
+    __________
+        geo_model : gp.core.model.Project
+            GemPy Model containing the faults.
+        fault : str
+            String or list of strings containing the name of faults to be clipped, e.g. ``faults='Fault1'``.
+        which : str, default: ``'first'``
+            Parameter to decide which end of the faults to clip. Options include ``'first'``, ``'last'``, or both,
+            e.g. ``'which='first'``.
+        buffer_first : Union[int, float]
+            Int or float value or list of values to clip the fault/s behind the first interface point,
+            e.g. ``'buffer_first=500'``.
+        buffer_last : Union[int, float]
+            Int or float value or list of values to clip the fault/s behind the last interface point,
+            e.g. ``'buffer_last=500'``.
+        i_size: Union[int, float]
+            Size of the plane in the i direction.
+        j_size: Union[int, float]
+            Size of the plane in the j direction.
+        invert_first : bool, default: ``'True'``
+            Invert clipping for first plane.
+        invert_last : bool, default: ``'False'``
+            Invert clipping for second plane.
+
+    Returns
+    _______
+        pv.core.pointset.PolyData
+            Clipped faults.
+
+    .. versionadded :: 1.1
+
+    See also
+    ________
+        create_plane_from_interface_and_orientation : Create PyVista plane from GemPy interface and orientations
+        DataFrames.
+        translate_clipping_plane : Translate clipping plane.
+
+    Example
+    _______
+
+
+
+    """
+    # Trying to import gempy but returning error if gempy is not installed
+    try:
+        import gempy as gp
+    except ModuleNotFoundError:
+        raise ModuleNotFoundError(
+            'GemPy package is not installed. Use pip install gempy to install the latest version')
+
+    # Checking that the fault is provided as string
+    if not isinstance(fault, str):
+        raise TypeError('Faults must be provided as one string for one fault ')
+
+    # Checking that the fault is a fault of the geo_model
+    if isinstance(fault, str):
+        if not fault in geo_model.surfaces.df['surface'][geo_model.surfaces.df['isFault'] == True].tolist():
+            raise ValueError('Fault is not part of the GemPy geo_model')
+
+        # Getting the fault DataFrames
+        fault_df_interfaces = gp.get_data(geo_model, 'surface_points')[
+            gp.get_data(geo_model, 'surface_points')['surface'] == fault].reset_index(drop=True)
+        fault_df_orientations = gp.get_data(geo_model, 'orientations', numeric=True)[
[…]
