+++ docs/getting_started/tutorial/index.rst
@@ -106,3 +106,4 @@ Each set of functions of GemGIS is collected in a different module. The function
     65_displaying_seismic_horizons_and_faults
     66_generating_voronoi_polygons
     67_rotating_gempy_input_data
+    68_Creating_Finite_Faults_with_GemGIS
