+++ gemgis/__init__.py
@@ -18,17 +18,15 @@ __correspondence_email__ = 'alexander.juestel@rwth-aachen.de'
 
 __affiliations__ = 'CGRE - RWTH Aachen University'
 
-__version_date__ = '2023-07-31'
+__version_date__ = '2023-08-18'
 
-__version__ = '1.1'
+__version__ = '1.0.12'
 
 __changelog__ = """What is new in version 1.1.:
 
 
 
 """
-
-
 
 from gemgis.gemgis import *
 import gemgis.vector as vector
