### 2023-08-29 #128 @NickleDave
Hi @AlexanderJuestel just checking back. I think we do have an editor that can run this review.

I know we were helping you with the API docs build but I'm a little out of the loop on where we left off. Any chance you can give me a rough estimate of when you might have a chance to get to that? I don't mean to rush you--I'm sure you have other obligations as well; I just don't want to lose track of this review.

My guess is that, if possible, switching to a pyproject.toml that includes your dependencies, and then using pure Python virtual environments on readthedocs, might make the build easier. But it's possible I'm missing the real issue.
