+++ gemgis/utils.py
@@ -2494,3 +2494,284 @@ def open_mpk(path_in: str):
 
     return dict_vector_data, dict_raster_data
 
+
+def convert_crs_seismic_data(path_in: str,
+                             path_out: str,
+                             crs_in: Union[str, pyproj.crs.crs.CRS],
+                             crs_out: Union[str, pyproj.crs.crs.CRS],
+                             cdpx: int = 181,
+                             cdpy: int = 185,
+                             vert_domain: str = 'TWT'):
+    """Convert CDP coordinates of seismic data to a new CRS.
+
+    Parameters
+    __________
+        path_in : str
+            Path to the original seismic data, e.g. ``path_in='seismic.sgy'``.
+        path_out : str
+            Path to the converted seismic data, e.g. ``path_out='seismic_converted.sgy'``.
+        crs_in : str, pyproj.crs.crs.CRS
+            Coordinate reference system of the original seismic data.
+        crs_out : str, pyproj.crs.crs.CRS
+            Coordinate reference system of the converted seismic data.
+        cdpx : int
+            Byte position for the X coordinates, default is ``cdpx=181``.
+        cdpy : int
+            Byte position for the Y coordinates, default is ``cdpx=185``.
+        vert_domain : str
+            Vertical sampling domain. Options include ``'TWT'`` and ``'DEPTH'``, default is ``vert_domain='TWT'``.
+
+    .. versionadded:: 1.1.1
+
+    """
+    # Trying to import segysak but returning error if segysak is not installed
+    try:
+        from segysak.segy import segy_loader, segy_writer
+    except ModuleNotFoundError:
+        raise ModuleNotFoundError(
+            'segysak package is not installed. Use pip install segysak to install the latest version')
+
+    # Checking that path_in is of type string
+    if not isinstance(path_in, str):
+        raise TypeError('path_in must be provided as string')
+
+    # Checking that path_out is of type string
+    if not isinstance(path_out, str):
+        raise TypeError('path_out must be provided as string')
+
+    # Checking that crs_in is of type string or pyproj CRS
+    if not isinstance(crs_in, (str, pyproj.crs.crs.CRS)):
+        raise TypeError('crs_in must be provided as string or pyproj CRS')
+
+    # Checking that crs_out is of type string or pyproj CRS
+    if not isinstance(crs_out, (str, pyproj.crs.crs.CRS)):
+        raise TypeError('crs_out must be provided as string or pyproj CRS')
+
+    # Checking that vert_domain is of type str
+    if not isinstance(vert_domain, str):
+        raise TypeError('vert_domain must be provided as string')
+
+    # Loading seismic data
+    seismic = segy_loader(path_in,
+                          vert_domain=vert_domain,
+                          cdpx=cdpx,
+                          cdpy=cdpy)
+
+    # Converting Seismic to DataFrame
+    df_seismic = seismic.to_dataframe()
+
+    # Checking that the CDP coordinates are in the DataFrame
+    if not {'cdp_x', 'cdp_y'}.issubset(df_seismic.columns):
+        raise ValueError(
+            'No coordinates found, please provide the byte positions where the X and Y data of the CDPs is stored')
+
+    # Extracting the length of the samples to reduce computing time
+    samples = len(df_seismic.index.get_level_values(1).unique())
+
+    # Resample DataFrame
+    df_seismic_resampled = df_seismic[::samples]
+
+    # Reprojecting Coordinates
+    df_seismic_reprojected = gpd.GeoDataFrame(geometry=gpd.points_from_xy(x=df_seismic_resampled['cdp_x'].values,
+                                                                          y=df_seismic_resampled['cdp_y'].values),
+                                              crs=crs_in).to_crs(crs_out)
+
[…]
