+++ docs/getting_started/tutorial/index.rst
@@ -107,3 +107,4 @@ Each set of functions of GemGIS is collected in a different module. The function
     66_generating_voronoi_polygons
     67_rotating_gempy_input_data
     68_Creating_Finite_Faults_with_GemGIS
+    70_Reprojecting_Seismic_Data
