+++ docs/getting_started/installation.rst
@@ -1,6 +1,6 @@
 .. _installation_ref:
 
-Installation (Tested 2023-01-01)
+Installation (Tested 2023-09-01)
 ===========================================================
 
 GemGIS is supported on Python version 3.10. Previous versions are officially not supported.
@@ -96,15 +96,15 @@ Installing GemGIS's dependencies manually
 
 Two of the main packages that GemGIS is dependent on are `rasterio <https://rasterio.readthedocs.io/en/latest/>`_ and `GeoPandas <https://geopandas.org/>`_. It is recommended to install these packages separately as they both depend on the `GDAL <https://gdal.org/>`_ translator library for raster and vector geospatial data. In addition, many smaller libraries like shaply or fiona will also be installed properly.
 
-Install the latest versions of GeoPandas and Rasterio (as of 2023-01-01). Please mind the quotation marks that are necessary when specifying the version numbers.::
+Install the latest versions of GeoPandas and Rasterio (as of 2023-09-01). Please mind the quotation marks that are necessary when specifying the version numbers.::
 
-   conda install -c conda-forge geopandas">=0.12.2" rasterio">=1.3.4"
+   conda install -c conda-forge geopandas">=0.13.2" rasterio">=1.3.8"
 
 .. image:: images/cmd6.png
 
 In order to visualize and work with meshes, the `PyVista <https://docs.pyvista.org/>`_ package is being installed::
 
-    conda install -c conda-forge pyvista">=0.37.0"
+    conda install -c conda-forge pyvista">=0.42."
 
 
 Installing GemGIS via PyPi
