+++ gemgis/__init__.py
@@ -18,9 +18,9 @@ __correspondence_email__ = 'alexander.juestel@rwth-aachen.de'
 
 __affiliations__ = 'CGRE - RWTH Aachen University'
 
-__version_date__ = '2023-08-18'
+__version_date__ = '2023-09-23'
 
-__version__ = '1.1.0'
+__version__ = '1.1.1'
 
 __changelog__ = """What is new in version 1.1.0:
 
