### 2023-10-04 e444cf4
EditHexagonFunctions
