+++ .readthedocs.yaml
@@ -5,6 +5,12 @@
 # Required
 version: 2
 
+# Set the version of Python
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.10"
+
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    builder: html
@@ -16,10 +22,7 @@ formats:
    - epub
    - pdf
 
-build:
-  os: ubuntu-22.04
-  tools:
-    python: 3.10
+
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
