+++ .readthedocs.yaml
@@ -26,7 +26,6 @@ formats:
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-   version: 3.8
    install:
       - method: pip
         path: .
