+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.10"
+    python: "mambaforge-22.9"
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
