+++ environment.yml
@@ -10,6 +10,4 @@ dependencies:
   - rasterio>=1.3.8
   # pyvista also install pooch and matplotlib
   - pyvista>=0.42.2
-  - pip
-  - pip:
-    - gemgis>=1.1
+  - gemgis>=1.1
