+++ environment_dev.yml
@@ -2,22 +2,17 @@ name: gemgis_dev_env
 channels:
   - conda-forge
 dependencies:
-  #- python>=3.10
-  #- geopandas=0.13.2
-  #- rasterio=1.3.8
-  #- pyvista=0.39.1
+  - python>=3.10
+  # geopandas will also install numpy, pandas, shapely, fiona, and pyproj
+  - geopandas>=0.14.0
+  # rasterio will also install affine
+  - rasterio>=1.3.8
+  # pyvista also install pooch and matplotlib
+  - pyvista>=0.42.2
+  - gemgis>=1.1
   - pip
   - pip:
-    - gempy==2.3
-    - rasterio==1.3.8
-    - geopandas==0.13.2
-    - pyvista
-    - shapely
-    - fiona
-    - pyproj
-    - gemgis
-    - matplotlib
-    - pandas
+    - gempy>=2.3
     - scipy
     - owslib
     - xarray
@@ -32,7 +27,6 @@ dependencies:
     - requests
     - PVGeo
     - pooch<1.5.0
-    - gemgis==1.0.11
     - pyqt5
     - nbsphinx
     - sphinx_markdown_tables
