+++ environment_dev.yml
@@ -10,31 +10,28 @@ dependencies:
   # pyvista also install pooch and matplotlib
   - pyvista>=0.42.2
   - gemgis>=1.1
-  - pip
-  - pip:
-    - gempy>=2.3
-    - scipy
-    - owslib
-    - xarray
-    - scikit-image
-    - rioxarray
-    - mplstereonet
-    - PyPDF2
-    - xmltodict
-    - geopy
-    - scikit-learn
-    - tqdm
-    - requests
-    - PVGeo
-    - pooch<1.5.0
-    - pyqt5
-    - nbsphinx
-    - sphinx_markdown_tables
-    - sphinx_gallery
-    - sphinx_rtd_theme
-    - sphinx_book_theme==0.3.3
-    - sphinx_copybutton
-    - sphinx_automodapi
-    - sphinx_numfig
-    - numpydoc
-    - build
+  - gempy>=2.3
+  - scipy
+  - owslib
+  - scikit-image
+  - rioxarray
+  - mplstereonet
+  - PyPDF2
+  - xmltodict
+  - geopy
+  - scikit-learn
+  - tqdm
+  - requests
+  - PVGeo
+  - pooch<1.5.0
+  - pyqt5
+  - nbsphinx
+  - sphinx_markdown_tables
+  - sphinx_gallery
+  - sphinx_rtd_theme
+  - sphinx_book_theme==0.3.3
+  - sphinx_copybutton
+  - sphinx_automodapi
+  - sphinx_numfig
+  - numpydoc
+  - build
