+++ environment_dev.yml
@@ -6,15 +6,16 @@ dependencies:
   # geopandas will also install numpy, pandas, shapely, fiona, and pyproj
   - geopandas>=0.14.0
   # rasterio will also install affine
-  - rasterio>=1.3.8
+  #- rasterio>=1.3.8 will be installed through rioxarray
   # pyvista also install pooch and matplotlib
-  - pyvista>=0.42.2
+  #- pyvista>=0.42.2 will be installed through pvgeo
   - gemgis>=1.1
   - gempy>=2.3
+  - rioxarray
+  - PVGeo
   - scipy
   - owslib
   - scikit-image
-  - rioxarray
   - mplstereonet
   - PyPDF2
   - xmltodict
@@ -22,7 +23,6 @@ dependencies:
   - scikit-learn
   - tqdm
   - requests
-  - PVGeo
   - pooch<1.5.0
   - pyqt5
   - nbsphinx
