+++ .github/workflows/pypi_release.yml
@@ -26,8 +26,9 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install setuptools wheel twine
-        pip install -r requirements.txt
+        #pip install setuptools wheel twine
+        #pip install -r requirements.txt
+        pip install gemgis[build]
     - name: Build and publish
       env:
         TWINE_USERNAME: ${{ secrets.PYPI_USERS }}
