+++ docs/conf.py
@@ -21,11 +21,11 @@ author = 'Alexander Juestel'
 # -- GemGIS configuration ---------------------------------------------------
 import os
 import sys
-sys.path.insert(0, os.path.abspath(".."))
-sys.path.insert(0, os.path.abspath('../..'))
-sys.path.insert(0, os.path.abspath("gemgis/"))
-sys.path.insert(0, os.path.abspath("../gemgis/"))
-sys.path.insert(0, os.path.abspath("../../gemgis/"))
+#sys.path.insert(0, os.path.abspath(".."))
+#sys.path.insert(0, os.path.abspath('../..'))
+#sys.path.insert(0, os.path.abspath("gemgis/"))
+#sys.path.insert(0, os.path.abspath("../gemgis/"))
+#sys.path.insert(0, os.path.abspath("../../gemgis/"))
 
 # The full version, including alpha/beta/rc tags
 
