+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = ["setuptools>=68.1.2",
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-package-dir = ["gemgis"]
+packages = ["gemgis"]
 
 [project]
 name = "GemGIS"
