+++ docs/getting_started/tutorial/index.rst
@@ -107,5 +107,6 @@ Each set of functions of GemGIS is collected in a different module. The function
     66_generating_voronoi_polygons
     67_rotating_gempy_input_data
     68_Creating_Finite_Faults_with_GemGIS
+    69_Export_GemPy_model_into_blender
     70_Reprojecting_Seismic_Data
     71_Opening_Rasters_From_OpenFileGDB
