+++ .github/workflows/pypi_release.yml
@@ -13,16 +13,14 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-
-    # Need the tags so that setuptools-scm can form a valid version number
-    - name: Fetch git tags
-      run: git fetch origin 'refs/tags/*:refs/tags/*'
-
-    - uses: actions/checkout@v2
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: '3.x'
+      - uses: actions/checkout@v2
+      # Need the tags so that setuptools-scm can form a valid version number
+      - name: Fetch git tags
+        run: git fetch origin 'refs/tags/*:refs/tags/*'
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.x'
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
