+++ docs/getting_started/gemgis.rst
@@ -1,0 +1,84 @@
+GemGIS API Reference
+====================
+
+The API reference provides an overview of all functions and methods implemented in GemGIS.
+
+gemgis.download\_gemgis\_data module
+------------------------------------
+
+.. automodule:: gemgis.download_gemgis_data
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.gemgis module
+--------------------
+
+.. automodule:: gemgis.gemgis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.misc module
+------------------
+
+.. automodule:: gemgis.misc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.postprocessing module
+----------------------------
+
+.. automodule:: gemgis.postprocessing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.raster module
+--------------------
+
+.. automodule:: gemgis.raster
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.utils module
+-------------------
+
+.. automodule:: gemgis.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.vector module
+--------------------
+
+.. automodule:: gemgis.vector
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.visualization module
+---------------------------
+
+.. automodule:: gemgis.visualization
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+gemgis.web module
+-----------------
+
+.. automodule:: gemgis.web
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: gemgis
+   :members:
+   :undoc-members:
+   :show-inheritance:
