+++ docs/getting_started/modules.rst
@@ -1,0 +1,7 @@
+GemGIS API Reference
+====================
+
+.. toctree::
+   :maxdepth: 4
+
+   gemgis
