+++ MANIFEST.in
@@ -5,4 +5,4 @@ include requirements_dev.txt
 
 recursive-exclude * __pycache__
 
-exclude .git/* .idea/* docs/* .github/* joss/* tests/*
+exclude .git/* .idea/* docs/** .github/* joss/** tests/*
