### 2023-11-23 ed71bd0
Fix setting of texture coordinates
