### 2023-11-26 #128 @AlexanderJuestel
@NickleDave Just to pick up the thread here again. The API Reference is working and the project was transferred to use a pyproject.toml file.

I may also have some capacity to work on some other issues prior to the start of the review if necessary to ensure that the review itself is smoothly :)
