+++ .github/workflows/testing.yml
@@ -1,41 +1,48 @@
-name: GemGIS CI
-
-on:
-  push:
-    branches: [ main, dev_aj, dev_alex ]
-
-  pull_request:
-    branches: [ main ]
-
-    # Allows you to run this workflow manually from the Actions tab
-    workflow_dispatch:
-
+name: Pytest
+on: [push]
 jobs:
-  # A workflow run is made up of one or more jobs that can run sequentially or in parallel
-  testing:
-    name: Testing
-    runs-on: "ubuntu-latest"
-    defaults:
-      run:
-        shell: bash -l {0}
+  run:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    env:
+      OS: ${{ matrix.os }}
+      PYTHON: '3.10'
     steps:
-      - uses: actions/checkout@v2
-      - uses: conda-incubator/setup-miniconda@v2
-        with:
-          activate-environment: gemgis_testing
-          environment-file: environment.yml
-          python-version: 3.9
-          channels: conda-forge
-      - name: List dependencies
-        run: |
-          conda list pygeos
-          conda list geopandas
-          conda list rasterio
-          conda list pyvista
-          conda list gemgis
-      - name: Install dependencies
-        run: |
-          pip install -r requirements_optional.txt
-      - name: Run tests with pytest
-        run: |
-          pytest
+    - uses: actions/checkout@v2
+    - name: Setup Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.10'
+    - name: Install Dependencies
+      run: |
+        pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Debug
+      run: |
+        ls -R  # List all files and directories recursively
+        pwd    # Print the current working directory
+    - name: Generate coverage report
+      working-directory: ./test
+      run: |
+        pip install pytest
+        pip install pytest-cov
+        pytest --cov=./ --cov-report=xml
+    - name: Debug
+      run: |
+        ls -R  # List all files and directories recursively
+        pwd    # Print the current working directory
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v3
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        directory: ./
+        #coverage/reports/
+        env_vars: OS,PYTHON
+        fail_ci_if_error: true
+        files: coverage.xml
+        #,./coverage2.xml,!./cache
+        flags: unittests
+        name: codecov-umbrella
+        verbose: true
