+++ .github/workflows/testing.yml
@@ -19,6 +19,7 @@ jobs:
       run: |
         pip install --upgrade pip
         pip install -r requirements.txt
+        pip install owslib geopy
     - name: Debug
       run: |
         ls -R  # List all files and directories recursively
