+++ .github/workflows/testing.yml
@@ -19,7 +19,7 @@ jobs:
       run: |
         pip install --upgrade pip
         pip install -r requirements.txt
-        pip install owslib geopy
+        pip install owslib geopy gempy==2.3.1 xmltodict PyPDF2 mplstereonet scikit-image scipy rioxarray
     - name: Debug
       run: |
         ls -R  # List all files and directories recursively
