+++ gemgis/misc.py
@@ -278,6 +278,9 @@ def get_meta_data(page: List[str]) -> list:
 
     .. versionadded:: 1.0.x
 
+    .. versionchanged:: 1.1.7
+    Adapting positions of coordinate values.
+
     Example
     _______
 
@@ -338,7 +341,7 @@ def get_meta_data(page: List[str]) -> list:
     well_number = well_number.split('Archiv-Nr.')[0]
 
     # Obtaining Depth of well
-    well_depth = page[page.index('Endteufe') + 2:page.index('Endteufe') + 3]
+    well_depth = page[page.index('Endteufe') + 3:page.index('Endteufe') + 4]
     well_depth = float(''.join(well_depth).replace(':', ''))
 
     # Obtaining Stratigraphie der Endteufe
@@ -355,20 +358,20 @@ def get_meta_data(page: List[str]) -> list:
     well_gemarkung = ''.join(well_gemarkung).replace(':', '')
 
     # Obtaining GK Coordinates of wells
-    well_coord_x_gk = page[page.index('Rechtswert/Hochwert') + 2:page.index('Rechtswert/Hochwert') + 3]
+    well_coord_x_gk = page[page.index('Rechtswert/Hochwert') + 3:page.index('Rechtswert/Hochwert') + 4]
     well_coord_x_gk = ''.join(well_coord_x_gk).replace(':', '')
 
-    well_coord_y_gk = page[page.index('Rechtswert/Hochwert') + 4:page.index('Rechtswert/Hochwert') + 5]
+    well_coord_y_gk = page[page.index('Rechtswert/Hochwert') + 5:page.index('Rechtswert/Hochwert') + 6]
     well_coord_y_gk = ''.join(well_coord_y_gk).replace(':', '')
 
     # Obtaining UTM Coordinates of wells
-    well_coord_x = page[page.index('East/North') + 2:page.index('East/North') + 3]
+    well_coord_x = page[page.index('East/North') + 3:page.index('East/North') + 4]
     well_coord_x = ''.join(well_coord_x).replace(':', '')
 
-    well_coord_y = page[page.index('East/North') + 4:page.index('East/North') + 5]
+    well_coord_y = page[page.index('East/North') + 5:page.index('East/North') + 6]
     well_coord_y = ''.join(well_coord_y).replace(':', '')
 
-    well_coord_z = page[page.index('Ansatzpunktes') + 2:page.index('Ansatzpunktes') + 3]
+    well_coord_z = page[page.index('Ansatzpunktes') + 3:page.index('Ansatzpunktes') + 4]
     well_coord_z = ''.join(well_coord_z).replace(':', '')
 
     # Obtaining Coordinates Precision
@@ -488,6 +491,9 @@ def get_meta_data_df(data: str,
 
     .. versionadded:: 1.0.x
 
+    .. versionchanged:: 1.1.7
+    Fixed bug in parsing PDF.
+
     Example
     _______
 
@@ -531,7 +537,7 @@ def get_meta_data_df(data: str,
     # Split Data
     data = data.split()
     data = '#'.join(data)
-    data = data.split('-Stammdaten')
+    data = data.split('-#Stammdaten')
     data = [item.split('|')[0] for item in data]
     data = [item.split('#') for item in data]
 
@@ -689,17 +695,17 @@ def get_stratigraphic_data(text: list,
     well_name = ''.join(well_name).replace(':', '')
 
     # Obtaining Depth of well
-    well_depth = text[text.index('Endteufe') + 2:text.index('Endteufe') + 3]
+    well_depth = text[text.index('Endteufe') + 3:text.index('Endteufe') + 4]
     well_depth = float(''.join(well_depth).replace(':', ''))
 
     # Obtaining UTM Coordinates of wells
[…]
