### 2023-12-22 46485e5
Ensure compat of create_depth_maps
