### 2024-01-09 #128 @NickleDave
Hi @AlexanderJuestel, could you clarify a little?

> I have not made any progress on the tests yet.

I see you have a set of tests but it looks like those tests might not be passing. Is that what you mean? I can't find where we discussed tests previously.

> I am getting into the final phase of my PhD

If you are about to very busy with finishing up your PhD--believe me, I sympathize--then it might be better to put the review on hold. We don't want you to feel overwhelmed!

We also need to make an effort to adhere to the [timelines](https://www.pyopensci.org/software-peer-review/our-process/review-timeline.html) that we specify in our review process, so that we're not putting too much demands on our editors and reviewers. If you don't think you could respond to reviews in anywhere near the two week timeline, then now may not be a good time to start a review. I will admit we probably are taking longer than two weeks for many reviews, but still it's not fair to editors or reviewers to have to keep a review on their to-do list for months, so we should try to minimize that.

Please say a little more about what's left to do with the tests, and whether you'll be able to meet the review timelines (roughly) given other things you have going on.
