### 2024-01-10 #128 @AlexanderJuestel
Hello @NickleDave,

we mentioned the tests briefly on Slack (see below). I do not have 100% coverage with the tests right now. There are quite a few missing. If that would not be a show-stopper for now, can also proceed with the review as you suggested in your last paragraph. If not, we should put the review on hold for now and continue later.

![image](https://github.com/pyOpenSci/software-submission/assets/45469915/fbb4f138-c42c-4f3c-ae69-1e8975513c17)
