### 2024-01-10 #128 @NickleDave
Thank you @AlexanderJuestel for clarifying, that helps.

For tests, the criterion to start review is just that you *have* a test suite and it runs in CI. You have met that criterion.

If I were a reviewer, I would definitely ask you to move towards full coverage as much as possible though.

Let's go ahead and start the review. You've worked hard to get GemGIS ready, and I sense that you want to take this over the finish line. I'll reach out to the editor.
