### 2024-02-29 #128 @yeelauren
## Editor response to review:
 :wave:  @AlexanderJuestel  hurrah! We've found reviewers.  I like the look of the number of tutorials and examples included thus far, looking forward to more feedback from our reviewers :)
---

## Editor comments
:wave: Hi @aleksandraradecka1 and @martinfleis! Thank you for volunteering to review
for pyOpenSci!

- [ ] reviewer 3 (if applicable)

Please get in touch with any questions or concerns! Your review is due: ~**March 29th**
---

Reviewers: @aleksandraradecka1 @martinfleis
Due date: **March 29th**
