### 2024-03-13 #128 @aleksandraradecka1
Hi Everyone,

as I already spoko with @yeelauren, my review will be short. I hope it will be helpful anyway.

First of all, congrats to @AlexanderJuestel and the Team on creating such an interesting and extensive library. What I find especially good is:
- consistent and intuitive naming of modules, functions, arguments etc. (the only exception seems to be https://gemgis.readthedocs.io/en/latest/getting_started/reference/vector_api/gemgis.vector.intersection_polygon_polygon.html which could be renamed to 'intersect_polygons' in order to keep all function names starting from a verb)
- clear functions' descriptions and examples (I would only check if each description is clear even when accessed as the first element of API reference, meaning without earlier contex e.g. https://gemgis.readthedocs.io/en/latest/getting_started/reference/raster_api/gemgis.raster.sample_interfaces.html)

Additonally, I think the library would benefit from improving the following elements:
- 'Overview' section on GitHub provides a much clearer explanation of the library's aim than the ''Welcome page' in the Documentation. I would consider standardizing this.
- Both in the 'Installation' and 'Tutorials and Basic Usage', part of the instructions provided place a lot of attention on things outside the scope of GemGIS functionality (e.g. Anaconda). I think such a part could be summarized in one or two sentences supplied with links to the respective resources.

Good luck with the further development of the library!
Aleksandra
