### 2024-04-28 #128 @SimonMolinsky
---

Hi @AlexanderJuestel

Please consider the part below a work in progress; I will check every tutorial in your package! This is not my evil nature, but I'm excited about GemGIS! It could be a very useful tool for my work. I am also considering using the package during Python geospatial workshops. I will let you know when my review is ready!

A.

- [ ] (Docs) What interpolation methods are used? - missing description
- [ ]

B. **Tutorials checks**
- [ ] Tutorial 1:
  - recurring imports of `gemgis` | `geopandas` | `matplotlib` (a grey area)
- [ ] Tutorial 2:
- [ ]

C. API checks

- [ ] `gg.vector.extract_xy()`

`drop_id`: Isn’t it too risky to drop columns without the user’s consent? In my opinion, dataframe cleaning should be the user’s responsibility. The method name focuses on extracting XY coordinates, and removing columns might be an unwanted surprise. Comparing the function to `gg.vector.extract_xy_linestring()` - this one does not remove columns.

- [ ]
