### 2024-05-26 #128 @SimonMolinsky
@AlexanderJuestel Thank you for your work. It was a pleasure testing `GemGIS`! I've checked all tutorials and the package itself, and here is the wrap-up of my notes (and I won't repeat the same problems stated by @martinfleis ). Not all points have the same weight, but for me as a reviewer, points 1, 2, 3, 5, 8, 9, 10, 11 are critical.

1. Required dependencies. I've seen that you assume that some high-level packages (e.g., `geopandas`) will install other packages (e.g., `fiona`), so there is no need to list the lower-level packages as required dependencies. I recommend listing all imported packages as the core dependencies, even if those are supposedly installed by listed packages.
2. Optional vs. required dependencies. Some dependencies enlisted as optional should be required if their absence means the process fails. Why? For example, I can install `GemGIS` and its core dependencies in the environment with Python 3.12, but some optional dependencies won't work in these settings (I couldn't install `GemPy` in an environment with Py==3.18).
3. Optional dependencies - installation. Please include a paragraph on how to install the full version of the package with all dependencies. It might be helpful for people who are not Python devs :)
4. Dependencies imports in tutorials: In some tutorials, dependencies are imported multiple times throughout code cells. I recommend doing all imports in the first cell; this way, a user will quickly find something that needs to be added. I've seen that in the later tutorials, you put all imports in the first cell. (Tutorial 1, 3)
5. Contribution: information about the installation of all dev dependencies.
6. (Docs) What interpolation methods are used? - missing description
7. `gg.vector.extract_xy()`
`drop_id`: Isn't dropping columns without the user's consent too risky? Dataframe cleaning should be the user's responsibility. The method name focuses on extracting XY coordinates and removing columns might be an unwanted surprise. Comparing the function to `gg.vector.extract_xy_linestring()` - this one does not remove columns. (When I tried more tutorials, I understood the reasoning behind this functionality; it's up to you what you want to do with this issue).
8. `gg.raster.calculate_difference()` | `gg.raster.resize_raster()` | `gg.raster.resize_by_array()` | `gg.raster.extract_contour_lines_from_raster()`

When input arrays have different sizes, I get Import Error: `ModuleNotFoundError: Scikit Image package is not installed. Use pip install scikit-image to install the latest version`

The case when *optional* dependency should be listed as **required**.
9. Tutorials & API - imports. Here, we have multiple issues, mainly related to imports. I recommend writing one paragraph on installing additional dependencies with links to these packages:

- Tutorial 14: `rioxarray` is required for `gg.visualization.read_raster()`.
- Tutorial 17: Info about `mplstereonet` installation should be included.
- Tutorial 18: `GemPy` installation.
- Tutorial 19: `owslib` package.
- Tutorial 21: `gg.web.load_as_files()` - `ModuleNotFoundError: tqdm package is not installed. Use pip install tqdm to install the latest version`
- Tutorial 22: `gempy` import
- Tutorial 23: a/a
- Tutorial 26: `gg.misc.load_pdf(path=file_path + 'test_data.pdf')` - `ModuleNotFoundError: PyPDF package is not installed. Use pip install pypdf to install the latest version`
- Tutorial 28: `gg.utils.load_surface_colors()` - `ModuleNotFoundError: xmltodict package is not installed. Use pip install xmltodict to install the latest version`
- Tutorial 31: `gg.utils.get_location_coordinate()` - `ModuleNotFoundError: GeoPy package is not installed. Use pip install geopy to install the latest version`
- Tutorial 35: `gg.visualization.add_row_to_boreholes()` - `ModuleNotFoundError: tqdm package is not installed. Use pip install tqdm to install the latest version`
- Tutorial 38: `bokeh` is required
- Tutorial 52: `gg.visualization.read_raster()` requires `rioxarray`.
- Tutorial 54: `gempy` import
- Tutorial 55: a/a
- Tutorial 56: `segysak` import
- Tutorial 57: `gempy` import
- Tutorial 61: `gempy` import
- Tutorial 62: `gg.raster.extract_contour_lines_from_raster()` error (`ModuleNotFoundError: No module named 'skimage")
- Tutorial 65: `segysak` import
- Tutorial 67: `gempy` import & `pv.set_jupyter_backend('client')` results in `ImportError: Please install `trame` and `ipywidgets` to use this feature.`
- Tutorial 68: `gempy` import
- Tutorial 69: missing `gemgis` import
- Tutorial 70: `segysak` import

10. Tutorials - missing data:

- Tutorial 15: data `file_path + '11_NI_sm.ts'` seems to be not downloaded.
- Tutorial 66: The tutorial's data is not available (it's not possible to download it)
- Tutorial 67: The tutorial's data does not exist.
- Tutorial 68: The tutorial's data does not exist.
- Tutorial 69: The tutorial's data does not exist.
- Tutorial 70: The tutorial's data does not exist.
- Tutorial 71: The tutorial's data is not provided.

11. Tutorials: misc - important

- Tutorial 5: Number of points, default `all` or `100`? **Why does a small number of points change the extent?**
- Tutorial 10: Got this error in cell 18:

"`python
p = pv.Plotter()

p.add_mesh(mesh=grid, scalars=grid["Elevation"], cmap='gist_earth')

p.show_grid(color='black')
p.set_background(color='white')
p.show()
```

"`text
KeyError: 'Data array (Elevation) not present in this dataset.'
```

because my version of PyVista returned a key with the name `Elevation [m]` (not `Elevation`). I have `PyVista==0.43.5`

The final view was inverted in my notebook:

![tutorial10-final-result](https://github.com/pyOpenSci/software-submission/assets/31246246/36f7e034-5f57-4e4f-b9b2-5f4bfbe42bcf)

- Tutorial 13: There is a section (Georeference Cross Section) which states:

```
The first step is to georeference the cross section. A tutorial on how to georeference maps is provided on this documentation page.
```

Shouldn't there be a link from the sentence `this documentation page`? I see screens taken from QGIS, but your tutorial (no 13) does not describe how to create cross-sections in QGIS.

- Tutorial 32: cell 11: `AttributeError: 'MultiBlock' object has no attribute 'active_scalars_name"
- Tutorial 41: cell 8 (or cell 3 if we preserve ordering):

```
Cell In[3], line 4
      1 import geopandas as gpd
      2 import fiona
----> 4 gpd.io.file.fiona.drvsupport.supported_drivers['KML'] = 'rw'
      6 layer_list = fiona.listlayers(file_path + 'KML_Samples.kml')
      7 layer_list

AttributeError: 'NoneType' object has no attribute 'drvsupport'
```
- Tutorial 42: don't know if [this view](https://gemgis.readthedocs.io/en/latest/_images/getting_started_tutorial_42_draping_linestrings_over_dem_in_pyvista_32_1.png) from `pyvista` is what you were looking for
- Tutorial 49: cell 21:

```
Cell In[28], line 1
----> 1 spline = pv.Spline(np.asarray(linestring),15)
      2 spline

File ~/miniforge3/envs/gemgis/lib/python3.12/site-packages/pyvista/core/utilities/parametric_objects.py:56, in Spline(points, n_points)
     18"  ""Create a spline from points.
     19
     20 Parameters
   (...)
     53
     54 """
     55 spline_function = _vtk.vtkParametricSpline()
---> 56 spline_function.SetPoints(pyvista.vtk_points(points, False))
     58 # get interpolation density
     59 u_res = n_points

File ~/miniforge3/envs/gemgis/lib/python3.12/site-packages/pyvista/core/utilities/points.py:48, in vtk_points(points, deep, force_float)
     46 # verify is numeric
     47 if not np.issubdtype(points.dtype, np.number):
---> 48     raise TypeError('Points must be a numeric type')
     50 if force_float:
     51     if not np.issubdtype(points.dtype, np.floating):

TypeError: Points must be a numeric type
```

- Tutorial 63: `gemgis` is not imported at the beginning (It is required in the 8th cell.) / Section **Combined Function** - function `show_well_log_along_well` was not defined.
- Tutorial 71: block [4] is:

"`python
gemgis.raster.read_raster_gdb(path=file_path + 'OpenFileGDB.gdb',
                              crs='EPSG:25832',
                              path_out=file_path)
```

But it should be:

"`python
gg.raster.read_raster_gdb(path=file_path + 'OpenFileGDB.gdb',
                                          crs='EPSG:25832',
                                          path_out=file_path)
```

12. Tutorials - misc - less important & optional

Tutorial 4: Maybe cropping by a polygon of irregular shape? This has nothing to do with the package's functionality or the tutorial's quality.
- Tutorial 11: Multiple empty cells.
- Tutorial 12: You have an error in the last cell of the tutorial. (But it works!)
- Tutorial 24: cell 12 - better if the output is hidden :)
- Tutorial 52: The first block with an example in the tutorial is commented on.

13. Linitng with PyCharm linter shows me multiple non-critical problems with your code. I recommended using a linter and checking if there are no hidden bugs in the code. (You might use `ruff` as Martin has done).
