### 2024-07-20 #128 @AlexanderJuestel
> Hi Everyone,
>
> [...]
>
> First of all, congrats to @AlexanderJuestel and the Team on creating such an interesting and extensive library. What I find especially good is:
>
> * consistent and intuitive naming of modules, functions, arguments etc. (the only exception seems to be https://gemgis.readthedocs.io/en/latest/getting_started/reference/vector_api/gemgis.vector.intersection_polygon_polygon.html which could be renamed to 'intersect_polygons' in order to keep all function names starting from a verb)

> [....]

The first point about renaming was addressed in https://github.com/cgre-aachen/gemgis/pull/336/commits/cbc9f31bc36358b832b3861face4395e9640e29b (mentioned in https://github.com/cgre-aachen/gemgis/issues/337)
