### 2024-07-20 #128 @AlexanderJuestel
@ aleksandraradecka1, could you please elaborate on what you meant by this comment? :)

> * clear functions' descriptions and examples (I would only check if each description is clear even when accessed as the first element of API reference, meaning without earlier contex e.g. https://gemgis.readthedocs.io/en/latest/getting_started/reference/raster_api/gemgis.raster.sample_interfaces.html)
