### 2025-07-29 #128 @lwasser
Hi everyone. Because this review has been open for over 2 years with no response since 2023, I am going to close it in its incomplete status. thank you everyone who put time into this review. pyOpenSci appreciates you ! ❤️
